Move RequestPath to parent server package
