+++ spring-test/src/main/java/org/springframework/mock/web/reactive/function/server/MockServerRequest.java
@@ -39,8 +39,8 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpRange;
 import org.springframework.http.MediaType;
-import org.springframework.http.server.reactive.PathContainer;
-import org.springframework.http.server.reactive.RequestPath;
+import org.springframework.http.server.PathContainer;
+import org.springframework.http.server.RequestPath;
 import org.springframework.http.server.reactive.ServerHttpRequest;
 import org.springframework.lang.Nullable;
 import org.springframework.util.Assert;
