+++ spring-web/src/main/java/org/springframework/http/server/reactive/AbstractServerHttpRequest.java
@@ -23,6 +23,7 @@ import java.util.regex.Pattern;
 
 import org.springframework.http.HttpCookie;
 import org.springframework.http.HttpHeaders;
+import org.springframework.http.server.RequestPath;
 import org.springframework.lang.Nullable;
 import org.springframework.util.CollectionUtils;
 import org.springframework.util.LinkedMultiValueMap;
