+++ spring-web/src/main/java/org/springframework/http/server/reactive/ServerHttpRequest.java
@@ -22,6 +22,7 @@ import org.springframework.http.HttpCookie;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpRequest;
 import org.springframework.http.ReactiveHttpInputMessage;
+import org.springframework.http.server.RequestPath;
 import org.springframework.lang.Nullable;
 import org.springframework.util.MultiValueMap;
