+++ spring-web/src/main/java/org/springframework/web/util/pattern/CaptureVariablePathElement.java
@@ -19,7 +19,7 @@ package org.springframework.web.util.pattern;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import org.springframework.http.server.reactive.PathContainer.UrlPathSegment;
+import org.springframework.http.server.PathContainer.UrlPathSegment;
 import org.springframework.lang.Nullable;
 
 /**
