+++ spring-web/src/main/java/org/springframework/web/util/pattern/WildcardPathElement.java
@@ -16,8 +16,8 @@
 
 package org.springframework.web.util.pattern;
 
-import org.springframework.http.server.reactive.PathContainer;
-import org.springframework.http.server.reactive.PathContainer.Element;
+import org.springframework.http.server.PathContainer;
+import org.springframework.http.server.PathContainer.Element;
 import org.springframework.web.util.pattern.PathPattern.MatchingContext;
 
 /**
