+++ spring-web/src/test/java/org/springframework/http/server/DefaultRequestPathTests.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.http.server.reactive;
+package org.springframework.http.server;
 
 import java.net.URI;
